+++ src/main/java/chess/util/JSonHandler.java
@@ -1,4 +1,4 @@
-package chess.service;
+package chess.util;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
