+++ src/main/resources/templates/chessBoardPage.hbs
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Corgi - chess</title>
-    <META HTTP-EQUIV="refresh" CONTENT="4">
+    <META HTTP-EQUIV="refresh" CONTENT="3">
     <link href="/style/board.css" rel="stylesheet" type="text/css">
 </head>
 <body>
