+++ src/main/resources/templates/mainPage.hbs
@@ -2,7 +2,7 @@
 <html lang="en">
 <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css" rel="stylesheet">
 <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap-theme.min.css" rel="stylesheet">
-<META HTTP-EQUIV="refresh" CONTENT="5">
+<META HTTP-EQUIV="refresh" CONTENT="4">
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/js/bootstrap.min.js"></script>
 <head>
     <meta charset="UTF-8">
