+++ src/test/java/chess/controller/ExceptionControllerTest.java
@@ -1,9 +1,8 @@
 package chess.controller;
 
-import chess.service.CookieHandler;
+import chess.util.CookieHandler;
 import chess.service.PlayerService;
 import chess.service.RoomService;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
